pyproject.toml
[project]
name = "z-butler"
requires-python = ">=3.12"
version = "3.18"
description = "Z butler bot for your server"
authors = [
  { name = "DroidZed", email = " <41507665+DroidZed@users.noreply.github.com>" },
]
license = "GPL-3.0-or-later"
readme = "README.md"
dependencies = [
  "httpx[http2] >= 0.27.2",
  "numpy >= 2.1.1",
  "beautifulsoup4 >= 4.12.3",
  "rich >= 13.8.1",
  "motor >= 3.6.0",
  "py-cord[speed] >= 2.6.1",
  "orjson >= 3.10.7",
  "pynacl >= 1.5.0",
  "msgspec >= 0.18.6",
  "aiofiles >= 24.1.0",
  "pillow >= 10.4.0",
  "aiohttp[speedups] >= 3.10.5",
  "python-decouple-typed >= 3.11.0",
  "coinpaprika-async >= 3.1.3",
  "loguru >= 0.7.2",
  "gtts >= 2.5.3",
  "typing-extensions >= 4.12.2",
]


[tool.uv]
default-groups = ["dev", "lint", "test"]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]

lint = [
    "ruff>=0.11.10",
]

test = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-httpx>=0.34.0",
]

[tool.pytest]
minversion = 6.0
testpaths = ["tests"]
asyncio_mode = "strict"
addopts = ["--import-mode=importlib", "-ra -q"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F403"]
